refactor(router-active): move to browser
