+++ src/platform/browser/router-active/index.ts
